RTL_sweep/requirements.txt
# RTL Sweep Pro — runtime dependencies
numpy>=1.24
scipy>=1.10
PyQt6>=6.5
pyqtgraph>=0.13
h5py>=3.9
matplotlib>=3.7        # used only for PNG exports of finished sweeps
pyrtlsdr>=0.3.0        # optional: real hardware. Mock fallback works without it.
